fixed rebase conflict with cargo.lock
