+++ sdk/ffi/shared/Cargo.toml
@@ -1,20 +1,16 @@
 [package]
-name = "nym-cpp-ffi"
-version = "0.1.1"
+name = "nym-ffi-shared"
+version = "0.1.0"
 edition = "2021"
 
-[lib]
-name = "nym_cpp_ffi"
-crate-type = ["cdylib"]
-
+# TODO change to load relative + remove this from the workspace exclude list
 [dependencies]
 # Async runtime
 tokio = { version = "1", features = ["full"] }
-# Nym clients, addressing, packet format, common tools (logging), ffi shared
+# Nym clients, addressing, packet format, common tools (logging)
 nym-sdk = { git = "https://github.com/nymtech/nym", branch = "master" }
 nym-bin-common = { git = "https://github.com/nymtech/nym", branch = "master" }
 nym-sphinx-anonymous-replies = { git = "https://github.com/nymtech/nym", branch = "master" }
-nym-ffi-shared = { path = "../shared" }
 # static var macro
 lazy_static = "1.4.0"
 # error handling
