+++ sdk/ffi/shared/README.md
@@ -1,0 +1,8 @@
+# FFI Shared ()s 
+This repo contains internal functions and runtime handling to be used by the C/C++ and Go FFI bindings. 
+
+## Usage
+Import the library as you would usually - see `../cpp/src/lib.rs` and `../go/src/lib.rs` for example usage. 
+
+
+
